conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_180_select=1.0=gpu
absl-py=0.2.2=py36_0
astor=0.6.2=py36_0
blas=1.0=mkl
bleach=1.5.0=py36_0
bzip2=1.0.6=h14c3975_5
ca-certificates=2018.03.07=0
cairo=1.14.12=h8948797_3
certifi=2018.11.29=py36_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.1.2=cuda9.0_0
cupti=9.0.176=0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gast=0.2.0=py36_0
glib=2.56.2=hd408876_0
graphite2=1.3.12=h23475e2_2
grpcio=1.12.1=py36hdbcaa40_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
html5lib=0.9999999=py36_0
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.3=0
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.2.0=hdf63c60_3
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.35=hbc83047_0
libprotobuf=3.5.2=h6f1eeef_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markdown=2.6.11=py36_0
mkl=2018.0.3=1
mkl_fft=1.0.1=py36h3010b51_0
mkl_random=1.0.1=py36h629b387_0
ncurses=6.1=hf484d3e_0
numpy=1.14.5=py36hcd700cb_3
numpy-base=1.14.5=py36hdbf6ddf_3
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1a=h7b6447c_0
pcre=8.42=h439df22_0
pip=10.0.1=py36_0
pixman=0.34.0=hceecf20_3
protobuf=3.5.2=py36hf484d3e_0
py-opencv=3.4.2=py36hb342d67_1
python=3.6.7=h0371630_0
readline=7.0=ha6073c6_4
setuptools=39.2.0=py36_0
six=1.11.0=py36h372c433_1
sqlite=3.25.3=h7b6447c_0
tensorboard=1.8.0=py36hf484d3e_0
tensorflow=1.8.0=hb11d968_0
tensorflow-base=1.8.0=py36hc1a7637_0
tensorflow-gpu=1.8.0=h7b35bdc_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
werkzeug=0.14.1=py36_0
wheel=0.31.1=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
